refactor(challenges): Simplify challenge form and dialog management
