+++ apps/nova/src/app/[locale]/(dashboard)/challenges/createChallengeDialog.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import ChallengeForm from './challengeForm';
+import ChallengeForm, { type ChallengeFormValues } from './challengeForm';
 import { Button } from '@tuturuuu/ui/button';
 import {
   Dialog,
@@ -10,11 +10,51 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@tuturuuu/ui/dialog';
+import { toast } from '@tuturuuu/ui/hooks/use-toast';
 import { Plus } from 'lucide-react';
+import { useRouter } from 'next/navigation';
+import { useState } from 'react';
 
 export default function CreateChallengeDialog() {
+  const router = useRouter();
+  const [open, setOpen] = useState(false);
+
+  const onSubmit = async (values: ChallengeFormValues) => {
+    try {
+      const url = '/api/v1/challenges';
+      const method = 'POST';
+
+      const response = await fetch(url, {
+        method,
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(values),
+      });
+
+      if (!response.ok) {
+        throw new Error('Failed to save challenge');
+      }
+
+      toast({
+        title: `Challenge created successfully`,
+        variant: 'default',
+      });
+
+      setOpen(false);
+      router.refresh();
+    } catch (error) {
+      console.error('Error saving challenge:', error);
+      toast({
+        title: 'Failed to save challenge',
+        description: error instanceof Error ? error.message : 'Unknown error',
+        variant: 'destructive',
+      });
+    }
+  };
+
   return (
-    <Dialog>
+    <Dialog open={open} onOpenChange={setOpen}>
       <DialogTrigger asChild>
         <Button className="gap-2">
           <Plus className="h-4 w-4" />
@@ -29,7 +69,7 @@ export default function CreateChallengeDialog() {
             with.
           </DialogDescription>
         </DialogHeader>
-        <ChallengeForm />
+        <ChallengeForm onSubmit={onSubmit} />
       </DialogContent>
     </Dialog>
   );
